Refactor utils.py
